+++ src/test/utils/options/help.test.ts
@@ -1,15 +1,10 @@
 import { ok } from "assert";
 
-import {
-    Options,
-    Logger,
-    ParameterType,
-    ParameterHint,
-} from "../../../lib/utils";
+import { Options, ParameterType, ParameterHint } from "../../../lib/utils";
 import { getOptionsHelp } from "../../../lib/utils/options/help";
 
 describe("Options - help", () => {
-    const options = new Options(new Logger());
+    const options = new Options();
     for (const decl of [
         { name: "td-option", help: "help", type: ParameterType.String },
         { name: "td-option2", help: "help" },